Further extracting of shared logic
