+++ shared/shared_utils.py
